backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain
langchain-groq
langchain-openai
opik
pydantic
python-dotenv
